Add Some Design Pattern

+++ backend/src/services/genre.service.ts
@@ -1,9 +1,9 @@
 import { CreateGenreDTO, UpdateGenreDTO } from "@/dtos/genre.dto";
 import { Genre } from "@/generated/prisma";
-import GenreRepository from "@/repositories/genre.repository";
+import { IGenreRepository } from "@/repositories/genre/IGenreRepository";
 
 class GenreService {
-  constructor(private readonly genreRepository: GenreRepository) {}
+  constructor(private readonly genreRepository: IGenreRepository) {}
 
   async createGenre(genreDto: CreateGenreDTO): Promise<Genre> {
     return this.genreRepository.create(genreDto);
@@ -22,14 +22,22 @@ class GenreService {
   }
 
   async updateGenre(id: string, updateDto: UpdateGenreDTO): Promise<Genre> {
+    const existingGenre = await this.genreRepository.findById(id);
+    if (!existingGenre) throw new Error(`Genre with ID ${id} not found.`);
+
     const updated = await this.genreRepository.update(id, updateDto);
     if (!updated) throw new Error(`Failed to update genre with ID ${id}`);
+
     return updated;
   }
 
   async deleteGenre(id: string): Promise<Genre> {
+    const existingGenre = await this.genreRepository.findById(id);
+    if (!existingGenre) throw new Error(`Genre with ID ${id} not found.`);
+
     const deleted = await this.genreRepository.delete(id);
     if (!deleted) throw new Error(`Failed to delete genre with ID ${id}`);
+    
     return deleted;
   }
 }
